robustcheckout: coerce exceptions to `str` and then `pycompat.bytestr` before displaying to screen (Bug 1723985) r=mhentges
Some of the robustcheckout code paths that throw exceptions assume the
exception can use `%s` splatting to format the exception in `ui.write`.
This was true under Python 2 but under Python 3, where Mercurial requires
byte strings, this is not the case. This commit changes the formatting code
to pass the exception through `str` to get a human-readable representation,
then through `pycompat.bytestr` which converts this representation into
a byte string.

Differential Revision: https://phabricator.services.mozilla.com/D124821

--HG--
extra : moz-landing-system : lando