vcsreplicator: limit number of heads printed in logs (Bug 1730708) r=mhentges
Previously we printed the full 40-char hash of each head, resulting in
large log entries on hgssh1. For repositories like try this is quite a
large log entry, which is added to various logs on each push to try.
This commit causes only the short hash to be printed to logs, and limits
the number of heads to 25.

Differential Revision: https://phabricator.services.mozilla.com/D125612

--HG--
extra : moz-landing-system : lando