testing: build test images from base centos (Bug 1740049) r=zeid
Deprecates the `secure image` builder functionality. We remove the
custom build context code and instead use the v-c-t root as the build
context, passing the path to the Dockerfile to specify which image
to build.

Since the custom builder was responsible for creating a path to
`ansible-playbook` in `/usr/bin`, we instead use the `ansible-playbook`
from `PATH` to run the image creation process.

Differential Revision: https://phabricator.services.mozilla.com/D130645

--HG--
extra : moz-landing-system : lando