testing: build hgcluster docker images in parallel
Specify `--parallel` on the docker-compose command line to make
image builds in parallel, speeding up build times considerably.
Since the parallel option is not aware that `hgweb0` and `hgweb1`
will use the same image on run, specify which imags to build on
the command line and omit `hgweb1` to avoid an unnecessary build.