configwizard: update evolve at the start of `configwizard` if already installed (Bug 1742622) r=mhentges
Right now configwizard has a single "evolve" step that informs the user
about evolve, clones, enables, and updates the extensions when a prompt
is accepted. This means for a user to update a configwizard-managed copy
of the evolve extension they must accept the configwizard prompt and go
through the entire wizard start to finish. Many users don't do this often
and skip the wizard after v-c-t updates.

This commit changes the wizard to update evolve before displaying any prompts
or dialogues. We only perform the update to evolve if the wizard is managing
the extensions. This behaviour should update evolve to the latest release
automatically whenever configwizard is run. Specifically this should update
evolve in the background when `./mach bootstrap` is run from a central clone.

Differential Revision: https://phabricator.services.mozilla.com/D131919

--HG--
extra : moz-landing-system : lando