clonebundles: parse gcp IP ranges from `cloud.json` instead of DNS (Bug 1749810) r=mhentges
This commit moves the GCP clonebundles preference code from using
DNS and assuming all requests are coming from `us-central1`, to
using `cloud.json` and serving a region-specific bundle where
appropriate.

We remove the recursive DNS query code from the IP scraper and
replace it with a copy of the AWS IP scraper function, with GCP
specific variables, JSON keys and IP object schema.

We change the `hgmo` extensions clonebundle manifest parsing
code to read the object as a JSON file conforming to the schema
and pass the parsed region value as the filter, instead of the
hard-coded `us-central1`. We update the `hgrc` template
to reflect the new file being `.json` instead of `.txt`, and
replace the old format test file with a sample `cloud.json`.
We also update the GCP IP address in `test-clonebundles.t`
to an IP within a network in the test file.

Since the DNS code has been removed, `dnspython` is removed as
a dependency of `requirements-tools-py3`.

We add a new `voluptuous` checker to assert all required GCP regions
are present in the JSON file. At this time this only includes the
existing `us-central1`, with more regions to be added in a separate
bug.

Differential Revision: https://phabricator.services.mozilla.com/D135904

--HG--
extra : moz-landing-system : lando