pushlog: move db commit and close to different stages of Mercurial transaction (Bug 1735633) r=zeid
Moves sqlite `commit` and `close` calls to different Mercurial transaction
hooks. Moving `commmit` to the `finalize` step should cause pushlog database
failures to reject the push without committing the Mercurial transaction.
Interestingly there is a Mercurial extension in the core hg repo that implements
sqlite storage for Mercurial and this extension uses `addfinalize` as the
transaction commit callback.

We also add `rollback` to the `abort` Mercurial transaction step as we
simply close the connection at the moment.

Differential Revision: https://phabricator.services.mozilla.com/D140157

--HG--
extra : moz-landing-system : lando