httpd: redirect requests for m-c tip's version of the Geckodriver `Cargo.toml` to Github (Bug 1876226) r=zeid
While Geckodriver is developed in m-c, it is released from Github
and the Github source is the most up to date stable version. Redirect
requests for the m-c `tip` copy of `Cargo.toml` to the Github copy
instead. This will reduce the load on hgmo caused by Geckodriver
projects in external CI pipelines, and give those pipelines a more
correct version of the file.

Differential Revision: https://phabricator.services.mozilla.com/D199452

--HG--
extra : moz-landing-system : lando