ansible: add configs to deploy Kafka to a dedicated cluster (Bug 1883093) r=zeid
We would like to migrate the Kafka processes off of hgweb/hgssh
and onto a dedicated cluster of nodes. This commit implements the
first phase required for the migration. First we add the new Kafka
nodes to the hosts file and add the DNS entries for the nodes to
the `zk_servers` mapping. Then we add a new `deploy-kafka.yml`
playbook and import it in the `deploy-hgmo.yml` playbook, conditionally
on the presence of a `skip_kafka` variable. We hook the `skip_kafka`
variable up to the `./deploy` command in the root of the repo in the
same fashion as the other `--skip-<hosts>` flags.

This patch will land, and we will deploy Kafka to the new nodes and do
some testing. At this stage we will still have Kafka running on the
hg nodes and can assert that the new cluster is sufficiently spec'd
for our purposes. At a later time we will remove Kafka from the
hg nodes.

Differential Revision: https://phabricator.services.mozilla.com/D204001

--HG--
extra : moz-landing-system : lando