vcsreplicator: add `--start-from` flag to `pushnotifications.py` (Bug 1885935) r=jcristau
Add support for a `--start-from` flag to the generic push notification
CLI. This flag can be used to set the consumer to a specific location
in the replication queue. Combined with `--skip`, this flag will
effectively move the consumer forward in the queue, read a message,
and commit the message. Subsequent calls to `get_message` by
consumers will return the latest messages in the queue.

Differential Revision: https://phabricator.services.mozilla.com/D204942

--HG--
extra : moz-landing-system : lando