hooks: make `prevent_symlinks.py` more efficient and correct (Bug 1886069) r=glandium
Switch to using `ctx.flags(path)` to locate the symlink flag
in a commit. This will cause the manifest to only be searched
once, whereas we are currently checking for the file presence
first and then searching again to find the flags.

Update the check that asserts a change is a symlink to look for
the presence of the `l` flag instead of checking that the returned
value is exactly the `l` flag. In theory multiple flags could be
returned, and thus a symlink would be missed.

Differential Revision: https://phabricator.services.mozilla.com/D205016

--HG--
extra : moz-landing-system : lando