treeclosure: get `treestatus_base_url` from the config (Bug 1894257) r=zeid
Change `treestatus_base_url` from a hard-coded value to a value
pulled from the config. Set the previous value as the default
if no value is found in the configuration. This will have the effect
of keeping the current value as the default and allowing us to switch
to the new Lando-Treestatus by updating the global config on hgssh.
Register the new config option in the `hgmo` extension to avoid a
warning in the tests, since the hook is implemented as an old-style
hook.

Differential Revision: https://phabricator.services.mozilla.com/D209019

--HG--
extra : moz-landing-system : lando