bundles: always upload new bundles instead of resetting bucket lifecycle (Bug 1896382) r=jcristau
The `CopyObject` operation we were using to reset the lifetime of an
object in S3 has a 5GB limit for atomic operations. Since our bundles
seem to be above 5GB now, we can no longer use this method. Remove the
code that checks for the presence of a key and runs `CopyObject` and
instead always upload a new bundle file to reset the object lifetime,
while we look for an alternative way to solve this problem.

Differential Revision: https://phabricator.services.mozilla.com/D210256

--HG--
extra : moz-landing-system : lando