ansible: re-add `kafka-broker` role to hgweb/hgssh (Bug 1883093) r=jcristau
The `kafka-broker` role was removed from hgweb/hgssh for the migration
of Kafka to a separate cluster. This change regressed the test suite
since we implicitly installed `kafka-broker` as part of `hg-web` instead
of in `test-kafkabroker`. Re-add the `kafka-broker` role to `hg-web`
and `hg-ssh` since the migration hasn't been fully completed, and Kafka
is still on those nodes. This brings the configs back in line with reality
and fixes the regression in the test suite.

Differential Revision: https://phabricator.services.mozilla.com/D211556

--HG--
extra : moz-landing-system : lando