kafka-broker: cleanup logs older than 7 days (Bug 1917814) r=jcristau
We rarely, if ever, look at the Kafka logs on any of the Kafka brokers.
If we do look, it's usually during an incident. Given the low amount
of space on kafka3 at the moment, this patch updates the log cleanup
to remove all log files older than 7 days.

Differential Revision: https://phabricator.services.mozilla.com/D221746

--HG--
extra : moz-landing-system : lando