bundles: upload bundles to Azure Cloud Storage (Bug 1632495) r=jcristau,jmoss
This commit adds bundle upload to Azure Cloud storage for each
Azure region where we run Firefox CI. We add the `AZURE_REGIONS`
variable with details on each region URL, bucket and region name.
In the same fashion as GCP, we upload only stream clone bundles to
each region. Azure credentials are pulled from the environment
with the `EnvionmentCredential` object, which are passed to a
`BlobClient` for each bundle. If the bundle already exists in
the bucket, reset the object expiration time to 7 days. If the
bundle doesn't exist, upload it to Azure. For each uploaded
bundle, create an entry in the manifest file with the region
tagged using the `azureregion=` parameter. A later commit will
filter on this parameter to serve an appropriate bundle during
clonebundles manifest filtering.

Differential Revision: https://phabricator.services.mozilla.com/D228754

--HG--
extra : moz-landing-system : lando