hgmo: add clonebundles manifest filtering for Azure clonebundles (Bug 1632495) r=jcristau
Add clonebundles manifest filtering for Azure regions in the same fashion
as the GCP and AWS regions. Add a new `hgmo.azureippath` config option
which holds a path to the Azure IP blocks file. When the path is defined
and the `azureregion=` string can be found in the manifest, parse the
file and compare the `addressPrefixes` property for each service with
the source IP address of the request, skipping all IPv6 networks. If
the source IP is within any of the specified networks, return the
`region` property and filter the manifest to only include bundles from
that region.

Differential Revision: https://phabricator.services.mozilla.com/D229183

--HG--
extra : moz-landing-system : lando