bundles: reset Azure bundle expiration by setting object metadata (Bug 1934193) r=jmoss
If a bundle file already exists in Azure, we currently try and
set the immutability policy to reset object expiry. Additionally,
request the blob's properties to get the current metadata,
add a timestamp field to the metadata and set the metadata
on the blob in Azure. This should reset the blob's expiry
by updating the "last modified" date on the blob.

Differential Revision: https://phabricator.services.mozilla.com/D231980

--HG--
extra : moz-landing-system : lando