hooks: remove Lando required check (Bug 1853470) r=jcristau
Remove the `lando_required.py` check and the associated
test file. Remove references to the hook class and remove
references to hgrc config options from Ansible.

Since the Lando required check was the only thing using
the `sentry_dsn`, we remove it from the Ansible defaults
file. Since it was the only encrypted variable in that
file, we also remove it from the list of encrypted files
in `.sops.yaml`.

Differential Revision: https://phabricator.services.mozilla.com/D231978

--HG--
extra : moz-landing-system : lando