bundles: use `hg debugbundle` to generate bundle spec (bug 1951886) r=sheehan
Run `hg debugbundle` against each bundle to populate the manifest's
BUNDLESPEC.  Check it against expectations to ensure we don't
unknowingly add new requirements.  This ends up adding
`revlog-compression-zstd` (support for which was added in hg 5.0) to the
requirements for a subset of stream bundles, which was previously
missing.  Conversely, the `dotencode`, `fncache` and `store`
requirements are removed from the spec: per mercurial upstream these
abstractions are handled by mercurial's vfs layer when unbundling.

Differential Revision: https://phabricator.services.mozilla.com/D240472

--HG--
extra : moz-landing-system : lando