hgtemplates: add a link to equivalent Github commit for changeset (Bug 1887984) r=jcristau
Add a new `git_repo_url` config option which points to the
equivalent Git repo that `git_commit` extras are built from.
Use this config option and the `git_commit` data to build
a URL to the Github commit this Mercurial changeset was
synced from.

Add a `commitextra.py` extension that can be used to set
arbitrary commit extras while running `hg commit`, and
use it in the `test-git-link.t` test.

Differential Revision: https://phabricator.services.mozilla.com/D247450

--HG--
extra : moz-landing-system : lando